Fix/mesh obfuscation

+++ platform/android/sdk/src/com/ansca/corona/VideoActivity.java
@@ -1,25 +1,9 @@
 //////////////////////////////////////////////////////////////////////////////
 //
-// Copyright (C) 2018 Corona Labs Inc.
-// Contact: support@coronalabs.com
-//
 // This file is part of the Corona game engine.
-//
-// Commercial License Usage
-// Licensees holding valid commercial Corona licenses may use this file in
-// accordance with the commercial license agreement between you and 
-// Corona Labs Inc. For licensing terms and conditions please contact
-// support@coronalabs.com or visit https://coronalabs.com/com-license
-//
-// GNU General Public License Usage
-// Alternatively, this file may be used under the terms of the GNU General
-// Public license version 3. The license is as published by the Free Software
-// Foundation and appearing in the file LICENSE.GPL3 included in the packaging
-// of this file. Please review the following information to ensure the GNU 
-// General Public License requirements will
-// be met: https://www.gnu.org/licenses/gpl-3.0.html
-//
-// For overview and more information on licensing please refer to README.md
+// For overview and more information on licensing please refer to README.md 
+// Home page: https://github.com/coronalabs/corona
+// Contact: support@coronalabs.com
 //
 //////////////////////////////////////////////////////////////////////////////
 
@@ -41,7 +25,7 @@ public class VideoActivity extends android.app.Activity {
 	private CoronaVideoView fVideoView;
 	
 	/** Unique integer ID assigned to this activity by this application's MediaManager class. */
-	private int fVideoId;
+	private long fVideoId;
 
 	/** Set true if the activity's onPause() was called. Set false if currently active. */
 	private boolean fWasActivityPaused;
@@ -161,7 +145,7 @@ public class VideoActivity extends android.app.Activity {
 			fVideoView.setMediaController(new android.widget.MediaController(themedContextWrapper));
 		}
 		fVideoView.setVideoURIUsingCoronaProxy(android.net.Uri.parse(getIntent().getExtras().getString("video_uri")));
-		fVideoId = getIntent().getExtras().getInt("video_id");
+		fVideoId = getIntent().getExtras().getLong("video_id");
 		fWasActivityPaused = false;
 		fWasVideoSuspended = false;
 		fVideoView.start();
