+++ platform/android/sdk/src/com/ansca/corona/events/EventManager.java
@@ -1,25 +1,9 @@
 //////////////////////////////////////////////////////////////////////////////
 //
-// Copyright (C) 2018 Corona Labs Inc.
-// Contact: support@coronalabs.com
-//
 // This file is part of the Corona game engine.
-//
-// Commercial License Usage
-// Licensees holding valid commercial Corona licenses may use this file in
-// accordance with the commercial license agreement between you and 
-// Corona Labs Inc. For licensing terms and conditions please contact
-// support@coronalabs.com or visit https://coronalabs.com/com-license
-//
-// GNU General Public License Usage
-// Alternatively, this file may be used under the terms of the GNU General
-// Public license version 3. The license is as published by the Free Software
-// Foundation and appearing in the file LICENSE.GPL3 included in the packaging
-// of this file. Please review the following information to ensure the GNU 
-// General Public License requirements will
-// be met: https://www.gnu.org/licenses/gpl-3.0.html
-//
-// For overview and more information on licensing please refer to README.md
+// For overview and more information on licensing please refer to README.md 
+// Home page: https://github.com/coronalabs/corona
+// Contact: support@coronalabs.com
 //
 //////////////////////////////////////////////////////////////////////////////
 
@@ -54,17 +38,17 @@ public class EventManager {
 		myEvents.clear();
 	}
 	
-	public void loadSound( int id, String name )
+	public void loadSound( long id, String name )
 	{
 		myController.getMediaManager().loadSound( id, name);
 	}
 
-	public void loadEventSound( int id, String name )
+	public void loadEventSound( long id, String name )
 	{
 		myController.getMediaManager().loadEventSound( id, name );
 	}
 
-	public void playSound( int id, String soundName, boolean loop )
+	public void playSound( long id, String soundName, boolean loop )
 	{
 		MediaEvent e = new MediaEvent( myController, id, soundName, MediaEvent.Event.PlaySound );
 		
@@ -73,28 +57,28 @@ public class EventManager {
 		addEvent( e );
 	}
 
-	public void stopSound( int id )
+	public void stopSound( long id )
 	{
 		Event e = new MediaEvent( myController, id, MediaEvent.Event.StopSound );
 		
 		addEvent( e );
 	}
 
-	public void pauseSound( int id )
+	public void pauseSound( long id )
 	{
 		Event e = new MediaEvent( myController, id, MediaEvent.Event.PauseSound );
 		
 		addEvent( e );
 	}
 
-	public void resumeSound( int id )
+	public void resumeSound( long id )
 	{
 		Event e = new MediaEvent( myController, id, MediaEvent.Event.ResumeSound );
 		
 		addEvent( e );
 	}
 
-	public void playVideo( int id, String name )
+	public void playVideo( long id, String name )
 	{
 		Event e = new MediaEvent( myController, id, name, MediaEvent.Event.PlayVideo );
